Revert changes to MCMC convergence checking, fix json serialization
